audit of 'uchar uc=0x80; ulong ul; ul = uc << 24;' class of problems
git-svn-id: https://svn.bingwo.ca/repos/sg3_utils/trunk@166 6180dd3e-e324-4e3e-922d-17de1ae2f315

+++ lib/sg_lib.c
@@ -265,7 +265,7 @@ sg_get_sense_info_fld(const unsigned char * sensep, int sb_len,
     case 0x70:
     case 0x71:
         if (info_outp)
-            *info_outp = (sensep[3] << 24) + (sensep[4] << 16) +
+            *info_outp = ((unsigned int)sensep[3] << 24) + (sensep[4] << 16) +
                          (sensep[5] << 8) + sensep[6];
         return (sensep[0] & 0x80) ? 1 : 0;
     case 0x72:
